+++ src/Products/PloneMeeting/MeetingItem.py
@@ -4364,7 +4364,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
                                   context=self.REQUEST))),
                 u"<acronym><a onclick='event.preventDefault()' class='tooltipster-group-users deactivated' "
                 u"style='display: inline-block; padding: 0'"
-                u"href='#' data-group_id='{0}' data-base_url='{1}'><img src='{1}/group_users.png' /></a></acronym>"
+                u"href='#' data-group_ids='{0}' data-base_url='{1}'><img src='{1}/group_users.png' /></a></acronym>"
                 .format(real_group_id, portal_url)))
         return u', '.join(res)
 
@@ -4384,7 +4384,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
                 value = u"{0} <acronym><a onclick='event.preventDefault()' " \
                     u"class='tooltipster-group-users deactivated' " \
                     u"style='display: inline-block; padding: 0'" \
-                    u"href='#' data-group_id='{1}' data-base_url='{2}'>" \
+                    u"href='#' data-group_ids='{1}' data-base_url='{2}'>" \
                     u"<img src='{2}/group_users.png' /></a></acronym>".format(
                         adviser['name'] + (not adviser['optional'] and auto_advice or u''),
                         get_plone_group_id(adviser['id'], 'advisers'),